simple-monitor/agent/requirements.txt
# Dependências para o agente Python (Windows)
psutil>=5.9.0
pywin32>=306
websocket-client>=1.6.0
